déplacements du curseur de nouveau linéaire et pas de problème de transition d'une aire à l'autre + vérifier si pas de problèmes avec la vitesse du curseur et avec la rotation

+++ MiAmusing/Source/BaseMidiSender.cpp
@@ -80,7 +80,7 @@ void TimeLine::setPeriod(int m_period)
 
 void TimeLine::setMidiTime(int idx, int newTime, int m_noteNumber,float m_velocity)
 {
-	newTime =  round((double)newTime * (double)currentPeriod / (double)period);
+	newTime =  (int)round((double)newTime * (double)currentPeriod / (double)period);
 	while (newTime > currentPeriod)
 		newTime -= currentPeriod;
 	if (idx < maxSize)
